+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Fixed
+- Fixed a bug where the scheduler would not re-try updating podgroup status after failure
+
 ## [v0.9.1] - 20250-09-15
 
 ### Added